BIG CHANGE.
when indexing, check to see if text exists in correctedtext folders. If so, use it and set the fulltext flag to true.

Switched to JSON format intead of XML for all solr POSTs. Much smaller and faster. Simplified lots of code. Seems to have a noticeable speed increase in indexing performance (small test archive completed in about 1/2 the time). needs more tests to determing real performance improvements